compare_experiment_behaviors : improve output path, file name
When prompted either type the name of the output folder for plots, or leave this blank for a dialog box.
When prompted, enter the “base” name of the output files for the plots, including the extension that specifies the image file type. The program will add '_relBehaviorPlot' and '_relBehaviorRatios' to this, so for example “this_expt.eps” gives outputs ““this_expt_relBehaviorPlot.eps” and “this_expt_relBehaviorRatios.eps”