Output filename; motion threshold
If there are subgroups, the subgroup name is now appended to the “behavior_counts” (or whatever)  file name containing summary statistics of each behavior for each dataset.
Only one place in the code to enter higher-level folder name. This is saved in the “basePath” variable in behaviors_main().
Basic characterization per fish: moving or not; Update analysis_parameters.yaml – new line:
  motion_speed_threshold_mm_second: 10