"seeing" now refers to any body position
Presently, “seeing” is assessed as the angle between a fish’s heading vector and the head-head vector to the other fish, i.e. the relative orientation angle, being less than some value. (Present parameters: the cosine of the relative orientation angle should be > 0.5). This is easy to implement, since the relative orientation angles are already calculated. In get_90_deg_frames() in behavior_identification.py, this is just one line per fish.
I alter this to indicate that any body position of the other fish is “seen.”